backend/requirements.txt
gunicorn
django
django-rest-framework
requests
django-cors-headers